Revert "Align host addresses (#11384) (#11836)"
This reverts commit 3296c13ef2999e1fc8391f8fa61145971f9e2d9a.
