+++ programs/bpf_loader/test_elfs/noop.so
