+++ sdk/src/bpf_loader.rs
@@ -1,1 +1,1 @@
-crate::declare_id!("BPFLoader2111111111111111111111111111111111");
+crate::declare_id!("BPFLoader1111111111111111111111111111111111");
