--- sdk/src/bpf_loader_deprecated.rs
@@ -1,1 +1,0 @@
-crate::declare_id!("BPFLoader1111111111111111111111111111111111");
