+++ sdk/src/lib.rs
@@ -6,7 +6,6 @@ extern crate self as solana_sdk;
 pub mod account;
 pub mod account_utils;
 pub mod bpf_loader;
-pub mod bpf_loader_deprecated;
 pub mod clock;
 pub mod commitment_config;
 pub mod decode_error;
